requirements.txt
tensorflow==2.9.1
numpy==1.22.4
pillow==7.0.0
